requirements.txt
anyio==4.9.0
certifi==2025.4.26
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mysql-connector-python==9.3.0
python-dotenv==1.1.0
python-telegram-bot==22.0
# Editable install with no version control (shadowdaemon==0.1)
-e .
sniffio==1.3.1
typing_extensions==4.13.2
